Add -z option to limit the CPU pressure when build
This feature will check the CPU pressure once a new child is born or
ready to be build. The function will then determine if the test can be
run or not, based on the current CPU pressure change. We always allow
the first task to be run.

Because of the Linux /proc/pressure design, we may or may not get an
update of the cpu pressure when the function is being called. In the
case where the CPU pressure have not yet being updated, we will follow
the last valid decision until an further update.

We tested it with openssl, and find that this feature not only can help
us reduce the CPU pressure by about 10%, and thus avoid some CPU
contention, but also reduce the compile time by about 10%.

Signed-off-by: Zheng Qiu <liqiu2003@gmail.com>